# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "岚山区劳动争议调解区块链平台，后台采用golang开发，使用gin + gorm + gorbac + jwt开发",
        "title": "岚山区劳动争议调解区块链平台",
        "termsOfService": "http://hafrans.com",
        "contact": {
            "name": "Chuuka Ro (Hafrans)",
            "url": "http://hafrans.com/support",
            "email": "lvzh@hafrans.com"
        },
        "license": {},
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/auth/sms/captcha/request": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "message"
                ],
                "summary": "发送短信验证码，验证信息",
                "parameters": [
                    {
                        "description": "请求",
                        "name": "email",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vo.SMSCaptchaRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "正常业务处理",
                        "schema": {
                            "$ref": "#/definitions/vo.SMSCaptchaResponse"
                        }
                    },
                    "401": {
                        "description": "未验证",
                        "schema": {
                            "$ref": "#/definitions/vo.Common"
                        }
                    },
                    "422": {
                        "description": "表单绑定失败",
                        "schema": {
                            "$ref": "#/definitions/vo.Common"
                        }
                    },
                    "500": {
                        "description": "表单绑定失败",
                        "schema": {
                            "$ref": "#/definitions/vo.Common"
                        }
                    }
                }
            }
        },
...
        "/api/v1/comment/create": {
            "post": {
                "description": "由双方人员创建对质信息，管理员不干涉，但是可以进行添加",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "case",
                    "comment"
                ],
                "summary": "根据 CaseID 创建双方对质信息",
                "parameters": [
                    {
                        "description": "提交表单",
                        "name": "case",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/vo.Comment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/vo.CommonData"
                        }
                    },
                    "401": {
                        "description": "没有认证",
                        "schema": {
                            "$ref": "#/definitions/vo.Common"
                        }
                    },
                    "422": {
                        "description": "绑定失败",
                        "schema": {
                            "$ref": "#/definitions/vo.Common"
                        }
                    }
                }
            }
        },
        "/api/v1/comment/delete/:id": {
            "get": {
                "description": "根据comment 的id号码删除comment，只有管理员、部门人员、以及创建者可以操作",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "labor",
                    "comment"
                ],
                "summary": "根据comment 的id号码删除comment",
                "parameters": [
                    {
                        "type": "number",
                        "description": "comment id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "正常业务处理",
                        "schema": {
                            "$ref": "#/definitions/vo.CommonData"
                        }
                    },
                    "401": {
                        "description": "未验证",
                        "schema": {
                            "$ref": "#/definitions/vo.Common"
                        }
                    }
                }
            }
        },
...
        "/api/v1/upload": {
            "get": {
                "description": "上传资源 ，支持 pdf,doc,docx,pdfx,txt,mp3,zip,jpg,gif,jpeg,png格式",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "utils"
                ],
                "summary": "上传资源",
                "parameters": [
                    {
                        "type": "string",
                        "description": "要上传的资源",
                        "name": "file",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/vo.Common"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/vo.Common"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "vo.Comment": {
            "type": "object",
            "required": [
                "case_id",
                "content"
            ],
            "properties": {
                "case_id": {
                    "description": "关联案件号",
                    "type": "string"
                },
                "content": {
                    "description": "对质内容",
                    "type": "string"
                },
                "id": {
                    "description": "对质ID",
                    "type": "integer"
                },
                "submitter": {
                    "description": "提交者",
                    "type": "string"
                },
                "submitter_phone": {
                    "description": "提交者联系方式",
                    "type": "string"
                },
                "submitter_type": {
                    "description": "提交者类型",
                    "type": "integer"
                }
            }
        },
        "vo.Common": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "unauthorized"
                },
                "status": {
                    "type": "integer",
                    "example": 401
                },
                "timestamp": {
                    "type": "string",
                    "example": "2048-05-06 12:34:56"
                }
            }
        },
        "vo.CommonData": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string",
                    "example": "unauthorized"
                },
                "status": {
                    "type": "integer",
                    "example": 401
                },
                "timestamp": {
                    "type": "string",
                    "example": "2048-05-06 12:34:56"
                }
            }
        },
...
        "vo.SMSCaptchaRequest": {
            "type": "object",
            "required": [
                "captcha_challenge",
                "captcha_code",
                "captcha_time",
                "phone"
            ],
            "properties": {
                "captcha_challenge": {
                    "type": "string",
                    "example": "asbduiasdvasilvdiwlqdulisbdaiauldvil=="
                },
                "captcha_code": {
                    "type": "string",
                    "example": "123456"
                },
                "captcha_time": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "vo.SMSCaptchaResponse": {
            "type": "object",
            "properties": {
                "challenge_code": {
                    "type": "string"
                },
                "challenge_time": {
                    "type": "string"
                }
            }
        },
...
    }
}